moved warmup to coord
Coordinator will now request the first drifts instead. This solves the issue of empty streams not triggering the warmup timer in the workers Functions.

+++ src/main/java/datatypes/internals/InitCoordinator.java
@@ -3,21 +3,21 @@ package datatypes.internals;
 import datatypes.InternalStream;
 
 public class InitCoordinator extends InternalStream {
-    private int warmup;
+    private long warmup;
 
     public InitCoordinator() {
     }
 
-    public InitCoordinator(int warmup) {
+    public InitCoordinator(long warmup) {
         this.warmup = warmup;
     }
 
 
-    public int getWarmup() {
+    public long getWarmup() {
         return warmup;
     }
 
-    public void setWarmup(int warmup) {
+    public void setWarmup(long warmup) {
         this.warmup = warmup;
     }
